+++ RELEASE_NOTES.md
@@ -14,6 +14,7 @@
 - Hardened header decoding.
 - Rejected misaligned section offsets in VGF 0.4.3 and newer during header decoding.
 - Fixed `DataView` equality operator.
+- Aligned VGF public version header to release component version.
 
 ### Highlights